+++ src/shared/tdoc-commons/services/tdoc-sql-mytbdb-objectdetection.adapter.ts
@@ -1,5 +1,5 @@
 import {SqlQueryBuilder} from '@dps/mycms-commons/dist/search-commons/services/sql-query.builder';
-import {CommonSqlObjectDetectionAdapter} from './common-sql-object-detection.adapter';
+import {CommonSqlObjectDetectionAdapter} from '@dps/mycms-commons/dist/action-commons/actions/common-sql-object-detection.adapter';
 import {TourDocSqlMytbDbConfig} from './tdoc-sql-mytbdb.config';
 
 export class TourDocSqlMytbDbObjectDetectionAdapter extends CommonSqlObjectDetectionAdapter {